requirements.txt
# ==============================================================================
# Aura AI System — Unified Requirements
# Python 3.11+
# All versions pinned to prevent conflicts
# ==============================================================================

# ── Core ──────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── LLM / AI ──────────────────────────────────────────────────────────────────
google-generativeai>=0.7.0

# ── API Server ────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0

# ── Vision ────────────────────────────────────────────────────────────────────
# Use opencv-python (not headless) so cv2.imshow works locally
opencv-python>=4.9.0
numpy>=1.26.0,<2.0.0          # <2.0 keeps mediapipe + ultralytics happy
pillow>=10.3.0

# Object detection
ultralytics>=8.2.0             # YOLOv8 / YOLOv9 / YOLOv11

# Hand + face detection
mediapipe>=0.10.14

# ── PyTorch (CPU build — works everywhere, no CUDA driver required) ────────────
# If you have an NVIDIA GPU + CUDA 12.1 driver, replace these with:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0

# ── Speech & Audio ────────────────────────────────────────────────────────────
faster-whisper>=1.0.0          # STT — CTranslate2 backend, CPU-fast
sounddevice>=0.4.6             # Microphone input (also used by VAD listener)
scipy>=1.13.0                  # Required by sounddevice
edge-tts>=6.1.9                # TTS — async, high quality (requires internet)
pyttsx3>=2.90                  # TTS fallback — fully offline
pygame>=2.5.0                  # Audio playback for edge-tts output

# ── Robotics ─────────────────────────────────────────────────────────────────
pyserial>=3.5                  # Serial / UART communication with MCU

# ── Utilities ─────────────────────────────────────────────────────────────────
rapidfuzz>=3.9.0               # Local intent matching
pyyaml>=6.0.1                  # Config file support
psutil>=5.9.0                  # CPU / memory stats for SystemInfoSkill
pycaw>=20240210                # Windows audio volume control (VolumeSkill)
comtypes>=1.4.1                # Windows COM — required by pycaw

# ── Protobuf — PINNED to prevent 3-way conflict ───────────────────────────────
# google-generativeai, mediapipe, and ultralytics all require protobuf.
# 4.x satisfies all three; 5.x breaks mediapipe <=0.10.14.
protobuf>=4.25.0,<5.0.0

# ==============================================================================
# DEPENDENCY CONFLICT NOTES:
#
# numpy<2.0.0   → mediapipe 0.10.x requires numpy<2.0, ultralytics works fine
# protobuf<5.0  → mediapipe 0.10.x doesn't support protobuf 5.x yet
# opencv-python → do NOT also install opencv-python-headless in same env
# torch CPU     → swap to CUDA build if GPU is available (see above comment)
# ==============================================================================
